fix: parse OpenRouter reasoning details for OpenAI models
