+++ src/inspect_ai/model/_providers/openrouter.py
@@ -1,17 +1,18 @@
 import json
 from logging import getLogger
-from typing import Annotated, Any, Literal, Union, cast
+from typing import Any, cast
 
 from openai.types.chat import (
     ChatCompletion,
     ChatCompletionMessageParam,
 )
-from pydantic import BaseModel, Field, JsonValue, TypeAdapter, ValidationError
+from pydantic import JsonValue
 from typing_extensions import NotRequired, TypedDict, override
 
 from inspect_ai._util.content import ContentReasoning
 from inspect_ai._util.error import PrerequisiteError
 from inspect_ai._util.logger import warn_once
+from inspect_ai.model import _openrouter_reasoning
 from inspect_ai.model._chat_message import ChatMessage
 from inspect_ai.model._model import RetryDecision
 from inspect_ai.model._model_output import ChatCompletionChoice
@@ -32,6 +33,16 @@ from .openai_compatible import OpenAICompatibleAPI
 OPENROUTER_API_KEY = "OPENROUTER_API_KEY"
 
 logger = getLogger(__name__)
+
+OPENROUTER_REASONING_DETAILS_SIGNATURE = (
+    _openrouter_reasoning.OPENROUTER_REASONING_DETAILS_SIGNATURE
+)
+openrouter_reasoning_details_to_reasoning = (
+    _openrouter_reasoning.openrouter_reasoning_details_to_reasoning
+)
+reasoning_to_openrouter_reasoning_details = (
+    _openrouter_reasoning.reasoning_to_openrouter_reasoning_details
+)
 
 
 class ErrorResponse(TypedDict):
@@ -280,104 +291,3 @@ class OpenRouterAPI(OpenAICompatibleAPI):
                 params[EXTRA_BODY]["reasoning"] = reasoning
 
         return params
-
-
-OPENROUTER_REASONING_DETAILS_SIGNATURE = "reasoning-details://"
-
-
-class ReasoningDetailBase(BaseModel):
-    id: str | None = Field(default=None)
-    format: str | None = Field(default=None)
-    index: int | None = Field(default=None)
-
-
-class ReasoningDetailSummary(ReasoningDetailBase):
-    type: Literal["reasoning.summary"]
-    summary: str
-
-
-class ReasoningDetailEncrypted(ReasoningDetailBase):
-    type: Literal["reasoning.encrypted"]
-    data: str
-
-
-class ReasoningDetailText(ReasoningDetailBase):
-    type: Literal["reasoning.text"]
-    text: str
-    signature: str | None = Field(default=None)
-
-
-ReasoningDetail = Annotated[
-    Union[ReasoningDetailSummary, ReasoningDetailEncrypted, ReasoningDetailText],
-    Field(discriminator="type"),
-]
-
-
-# openrouter uses reasoning_details
-# https://openrouter.ai/docs/guides/best-practices/reasoning-tokens#responses-api-shape
-def openrouter_reasoning_details_to_reasoning(
-    reasoning_details: list[dict[str, Any]],
-) -> ContentReasoning:
-    # store the full data structure in the signature for replay
-    details_json = json.dumps(reasoning_details)
-    signature = f"{OPENROUTER_REASONING_DETAILS_SIGNATURE}{details_json}"
-
-    # attempt to parse out the details
-    try:
-        adapter = TypeAdapter(list[ReasoningDetail])
-        details = adapter.validate_python(reasoning_details)
-    except ValidationError as ex:
-        logger.warning(
-            f"Error parsing OpenRouter reasoning details: {ex}\n\n{details_json}"
-        )
-        return ContentReasoning(reasoning=details_json, signature=signature)
-
-    # collect reasoning fields from details
-    reasoning: str | None = None
-    summary: str | None = None
-    redacted: bool = False
-    for detail in details:
-        match detail.type:
-            case "reasoning.summary":
-                summary = detail.summary
-            case "reasoning.text":
-                reasoning = detail.text
-            case "reasoning.encrypted":
-                if reasoning is not None:
-                    summary = reasoning
-                reasoning = detail.data
-                redacted = True
-
-    # resolve reasoning
-    if reasoning is None:
-        # summary becomes reasoning if there is no reasoning
-        if summary is not None:
-            reasoning = summary
-            summary = None
-        # otherwise this an unepxected state
-        else:
-            logger.warning(
-                f"Error parsing OpenRouter reasoning details: Reasoning content not provided.\n\n{details_json}"
-            )
-            return ContentReasoning(reasoning=details_json, signature=signature)
-
-    # return reasoning
-    return ContentReasoning(
-        reasoning=reasoning, summary=summary, redacted=redacted, signature=signature
-    )
-
-
-def reasoning_to_openrouter_reasoning_details(
-    content: ContentReasoning,
-) -> dict[str, Any] | None:
-    if content.signature and content.signature.startswith(
-        OPENROUTER_REASONING_DETAILS_SIGNATURE
-    ):
-        return {
-            "reasoning_details": json.loads(
-                content.signature.replace(OPENROUTER_REASONING_DETAILS_SIGNATURE, "", 1)
-            )
-        }
-
-    # default to no handling
-    return None
